[refactor] 함수명 구체화
- 불필요 함수 제거
