+++ android/app/src/main/java/com/teamtripdraw/android/ui/common/bindingAdapter/ImageBindingAdapter.kt
@@ -1,6 +1,5 @@
 package com.teamtripdraw.android.ui.common.bindingAdapter
 
-import android.view.View
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
@@ -8,15 +7,14 @@ import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.stfalcon.imageviewer.StfalconImageViewer
 import com.teamtripdraw.android.R
-import java.io.File
 
-@BindingAdapter("app:setImageVisibility")
-fun ImageView.setImageVisibility(imgUrl: String?) {
-    if (imgUrl.isNullOrBlank()) {
-        this.visibility = View.GONE
-        return
-    }
-    this.visibility = View.VISIBLE
+@BindingAdapter("app:loadPostImage")
+fun ImageView.loadPostImage(imgUrl: String?) {
+    Glide.with(this.context)
+        .load(imgUrl)
+        .placeholder(R.drawable.shape_td_gray_fill_0_rect)
+        .error(R.drawable.img_default_image)
+        .into(this)
 }
 
 @BindingAdapter("app:clickToShowImageViewer")
@@ -32,8 +30,8 @@ fun ImageView.clickToShowImageViewer(imgUrl: String?) {
     }
 }
 
-@BindingAdapter("app:setRoundImage")
-fun ImageView.setRoundImage(imgUrl: String?) {
+@BindingAdapter("app:loadThumbnailRoundImage")
+fun ImageView.loadThumbnailRoundImage(imgUrl: String?) {
     Glide.with(this.context)
         .load(imgUrl)
         .placeholder(R.drawable.shape_td_gray_fill_0_rect)
@@ -42,27 +40,8 @@ fun ImageView.setRoundImage(imgUrl: String?) {
         .into(this)
 }
 
-@BindingAdapter("app:setRoundImageFile")
-fun ImageView.setRoundImageFile(imgFile: File?) {
-    Glide.with(this.context)
-        .load(imgFile)
-        .placeholder(R.drawable.shape_td_gray_fill_0_rect)
-        .error(R.drawable.shape_td_gray_fill_0_rect)
-        .transform(CenterCrop(), RoundedCorners(20))
-        .into(this)
-}
-
-@BindingAdapter("app:setFileVisibility")
-fun ImageView.setFileVisibility(imgFile: File?) {
-    if (imgFile == null) {
-        this.visibility = View.GONE
-        return
-    }
-    this.visibility = View.VISIBLE
-}
-
-@BindingAdapter("app:setTripThumbnailImage")
-fun ImageView.setTripThumbnailImage(imgUrl: String) {
+@BindingAdapter("app:loadTripThumbnailImage")
+fun ImageView.loadTripThumbnailImage(imgUrl: String) {
     Glide.with(this.context)
         .load(imgUrl)
         .placeholder(R.drawable.shape_td_gray_fill_0_rect)
