+++ android/app/src/main/res/layout/activity_history_detail.xml
@@ -53,7 +53,7 @@
                         app:layout_constraintEnd_toEndOf="parent"
                         app:layout_constraintStart_toStartOf="parent"
                         app:layout_constraintTop_toTopOf="parent"
-                        app:setImage="@{historyDetailViewModel.previewTrip.imageUrl}"
+                        app:loadPostImage="@{historyDetailViewModel.previewTrip.imageUrl}"
                         tools:background="@color/td_light_blue" />
 
                     <View
@@ -72,7 +72,7 @@
                         android:scaleType="centerCrop"
                         app:layout_constraintBottom_toBottomOf="@id/iv_history_detail_image"
                         app:layout_constraintEnd_toEndOf="@id/iv_history_detail_image"
-                        app:setImage="@{historyDetailViewModel.previewTrip.routeImageUrl}"
+                        app:loadPostImage="@{historyDetailViewModel.previewTrip.routeImageUrl}"
                         app:setImageVisibility="@{historyDetailViewModel.previewTrip.routeImageUrl}"
                         tools:background="@color/td_gray" />
