+++ android/app/src/main/res/layout/activity_post_writing.xml
@@ -139,8 +139,8 @@
             android:layout_height="56dp"
             app:layout_constraintBottom_toBottomOf="@+id/et_writing"
             app:layout_constraintEnd_toEndOf="@+id/et_writing"
-            app:setFileVisibility="@{postWritingViewModel.imageFile}"
-            app:setRoundImageFile="@{postWritingViewModel.imageFile}" />
+            app:loadThumbnailRoundImage="@{postWritingViewModel.imageFile}"
+            app:setImageVisibility="@{postWritingViewModel.imageFile}" />
 
         <com.google.android.material.divider.MaterialDivider
             android:id="@+id/divider_content_file"
