+++ android/app/src/main/res/layout/item_history_detail_posts.xml
@@ -27,8 +27,8 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
+            app:loadThumbnailRoundImage="@{postItem.thumbnail}"
             app:setImageVisibility="@{postItem.thumbnail}"
-            app:setRoundImage="@{postItem.thumbnail}"
             tools:background="@drawable/shape_td_gray_fill_12_rect" />
 
         <TextView
